test: move inline tests to integration test suites
