+++ tests/dae/main.rs
@@ -1,0 +1,1 @@
+mod from_fn;
