+++ tests/dde/main.rs
@@ -1,3 +1,4 @@
 mod accuracy;
+mod from_fn;
 mod interpolation;
 mod systems;
