+++ tests/ode/main.rs
@@ -1,5 +1,6 @@
 mod accuracy;
 mod comparison;
 mod errors;
+mod from_fn;
 mod interpolation;
 mod systems;
